Use newer package for von-agent
I've created a new dockerfile which I'm using to build an environment
from which I can write code that uses von-agent.  You run this
container with the repository mounted in as a volume and set
`--net vonconnector_default` so that agent code running from inside
the container will be able to reach the test nodes that were started
with the `./manage start_nodes` command.  This image also clones and
install the `von-agent` python module instead of using the dump that
already existed in the repository.  This should allow me to start
iterating on the von-connector code and begin building our  own agent
apps.