Use later versions of Indy packages
It seems that the PSPC group is using newer versions of the Indy
packages that the von-connector code we have been investigating thus
far.  In order to use the latest `von-agent` code, I've attempted to
update my dev image to use the versions of these packages that I
see in the PSPC repositories.  One weird thing was that I had to
force `python3-indy` to install `1.3.1` because the `von-agent==0.3.5`
installation kept pulling `1.3.0` even though that isn't the version
listed in `von-agent`'s `requirements.txt` file.