Update packages used for starting nodes
The PSPC nodes are using newer packages for their sandbox node pool,
so I thought it would be a good idea to update the node pool that
I'm developing with to their listed versions.  I also optimized the
docker-compose file so that `docker-compose build` doesn't try to
build the same node image four times in a row, which is a painfully
slow build.